Code/sourcecode/Cloud_function/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
gcsfs
fsspec
pandas
pandas-gbq
colorlog